update
done

+++ app/src/main/java/com/arhamsoft/deskilz/networking/networkModels/GetPlayerAccountData.kt
@@ -2,7 +2,7 @@ package com.arhamsoft.deskilz.networking.networkModels
 
 data class GetPlayerAccountData( //getcustomplayerdata and getgamecustomdata same model
     //    val userTickets: Int=0,
-    val userDollers: Double=0.0,
-    val userDeskillzCurrency: Double=0.0,
-    val busd: Double=0.0
+    val userDollers: Float=0.0f,
+    val userDeskillzCurrency: Float=0.0f,
+    val busd: Float=0.0f
 )
